Added Required Dependencies And Updated README
Soon, I might go in and change the README more to reflect the
dependencies that I added manually. I kept these in by accident but
maybe this can take out a step for the powerusers trying to compile this
from scratch.

Documentation Update

Updated the README with more information at the start.